requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements-txt --no-hashes -o requirements.txt
asgiref==3.11.1
    # via
    #   django
    #   django-cors-headers
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.2.25
    # via
    #   pyogrio
    #   pyproj
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.7
    # via pdfminer-six
colorama==0.4.6 ; sys_platform == 'win32'
    # via pytest
cryptography==46.0.7
    # via pdfminer-six
django==6.0.3
    # via
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
    #   guayacanes
django-cors-headers==4.9.0
    # via guayacanes
django-filter==25.2
    # via guayacanes
djangorestframework==3.17.1
    # via
    #   djangorestframework-gis
    #   drf-spectacular
    #   guayacanes
djangorestframework-gis==1.0
    # via guayacanes
drf-spectacular==0.29.0
    # via guayacanes
geopandas==1.1.3
    # via guayacanes
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
numpy==2.4.4
    # via
    #   geopandas
    #   pandas
    #   pyogrio
    #   shapely
packaging==26.0
    # via
    #   geopandas
    #   pyogrio
    #   pytest
pandas==3.0.2
    # via geopandas
pdfminer-six==20251230
    # via pdfplumber
pdfplumber==0.11.9
    # via guayacanes
pillow==12.1.1
    # via
    #   guayacanes
    #   pdfplumber
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.11
    # via guayacanes
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pygments==2.19.2
    # via pytest
pyogrio==0.12.1
    # via geopandas
pypdfium2==5.7.0
    # via pdfplumber
pyproj==3.7.2
    # via geopandas
pytest==9.0.2
    # via pytest-django
pytest-django==4.12.0
python-dateutil==2.9.0.post0
    # via pandas
python-decouple==3.8
    # via guayacanes
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
shapely==2.1.2
    # via geopandas
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
tzdata==2025.3 ; sys_platform == 'emscripten' or sys_platform == 'win32'
    # via
    #   django
    #   pandas
uritemplate==4.2.0
    # via drf-spectacular
